move test_sharding_transaction to test_sharding_vm.py
